fix(modifications_history): Reworked the migration logic for historical tables (#2809)
When the state rewind window of 7d was triggered, the
`is_migration_in_progress` was repeatedly called, resulting in multiple
iterations over the empty `ModificationsHistoryV1` table, causing weird
behaviour with rocksdb as noticed in the flamegraph via pyroscope:
https://github.com/FuelLabs/fuel-core/pull/2808

Iteration was slow because compaction didn't have a chance to clean up
V1 table. This change removes iteration from the migration process.

---------

Co-authored-by: green <xgreenx9999@gmail.com>
Co-authored-by: AurelienFT <aurelien.foucault@epitech.eu>